refactor(app-core): extract useEditorModeSync from RetainedDenebEditor
<RetainedDenebEditor> grew to own five distinct concerns:

  1. Render-latch (hasOpenedOnce, transition detection)
  2. Viewport-match gate (isPendingSettle, RAF/resize/timer)
  3. interface.type sync on entry
  4. ModalDialog cleanup on exit
  5. Deferred new-project ModalDialog auto-open

Concerns 3-5 are pure side-effect bridges to the global stores — no rendered output, no interaction with the gate machinery beyond reading isEditorMode / isPendingSettle / hasOpenedOnce. Lift them into useEditorModeSync, kept in a sibling file. The component is now ~50 lines shorter and focused on render decisions and the gate effect.

Behaviour is unchanged: same effects, same deps, same fired-once latch on the Create modal. Existing tests for the latch helper continue to pass.

Refs:
- maintainability review (five distinct responsibilities)
